+++ src/audiobuffer.h
@@ -19,17 +19,17 @@ typedef struct {
   volatile uint64_t frames;
 
   // all indexes and ofsets in pcm samples
-  volatile size_t available;
-  volatile size_t read_index;
-  volatile size_t write_index;
+  volatile uint32_t available;
+  volatile uint32_t read_index;
+  volatile uint32_t write_index;
 
   uv_mutex_t mutex;
 
   // read/write begin/end state
   uint8_t r_begin;
   uint8_t w_begin;
-  size_t r_available, r_max_samples, r_index, r_count;
-  size_t w_available, w_max_samples, w_index, w_count;
+  uint32_t r_available, r_max_samples, r_index, r_count;
+  uint32_t w_available, w_max_samples, w_index, w_count;
 
   float data[];
 } audiobuffer_t;
@@ -41,7 +41,7 @@ typedef struct {
  * @param frames Buffer capacity in PCM frames
  * @return Pointer to initialized buffer
  */
-audiobuffer_t* audiobuffer_create(audio_format_t af, size_t frames);
+audiobuffer_t* audiobuffer_create(audio_format_t af, uint32_t frames);
 
 /**
  * Free memory
@@ -85,7 +85,7 @@ void audiobuffer_set_frames(audiobuffer_t* b, uint64_t frame);
  * @param max_frames Maximum frames number that can be processed in this routine
  * @return Number of frames available for reading
  */
-size_t audiobuffer_read_begin(audiobuffer_t* b, const size_t max_frames);
+uint32_t audiobuffer_read_begin(audiobuffer_t* b, const uint32_t max_frames);
 
 /**
  * Prepare pointer with samples for reading
@@ -94,7 +94,7 @@ size_t audiobuffer_read_begin(audiobuffer_t* b, const size_t max_frames);
  * @param ptr Returned pointer to data
  * @return Number of frames that can be read from provided pointer
  */
-size_t audiobuffer_read(audiobuffer_t* b, float** ptr);
+uint32_t audiobuffer_read(audiobuffer_t* b, float** ptr);
 
 /**
  * Mark specified frames as consumed
@@ -103,7 +103,7 @@ size_t audiobuffer_read(audiobuffer_t* b, float** ptr);
  * @param frames Frames that will be marked as consumed
  * @return Number of marked frames from beginning of current read routine
  */
-size_t audiobuffer_read_consume(audiobuffer_t* b, const size_t frames);
+uint32_t audiobuffer_read_consume(audiobuffer_t* b, const uint32_t frames);
 
 /**
  * End current read routine
@@ -113,7 +113,7 @@ size_t audiobuffer_read_consume(audiobuffer_t* b, const size_t frames);
  * @param b Pointer to initialized buffer
  * @return Number of frames was proccessed by current read routine
  */
-size_t audiobuffer_read_end(audiobuffer_t* b);
+uint32_t audiobuffer_read_end(audiobuffer_t* b);
 
 /**
  * Begin writing routine
@@ -124,7 +124,7 @@ size_t audiobuffer_read_end(audiobuffer_t* b);
  * @param max_frames Maximum number of frames number that can be processed in this routine
  * @return Free space available for writing in frames 
  */
-size_t audiobuffer_write_begin(audiobuffer_t* b, const size_t max_frames);
+uint32_t audiobuffer_write_begin(audiobuffer_t* b, const uint32_t max_frames);
 
 /**
  * Prepare pointer for writing samples
@@ -133,7 +133,7 @@ size_t audiobuffer_write_begin(audiobuffer_t* b, const size_t max_frames);
  * @param ptr Returned pointer to data
  * @return Number of frames that can be written to the provided pointer
  */
-size_t audiobuffer_write(audiobuffer_t* b, float** ptr);
+uint32_t audiobuffer_write(audiobuffer_t* b, float** ptr);
 
 /**
  * Mark specified frames as filled
@@ -142,7 +142,7 @@ size_t audiobuffer_write(audiobuffer_t* b, float** ptr);
  * @param frames Frames that will be marked as filled
  * @return Number of marked frames from beginning of current write routine
  */
-size_t audiobuffer_write_fill(audiobuffer_t* b, const size_t frames);
+uint32_t audiobuffer_write_fill(audiobuffer_t* b, const uint32_t frames);
 
 /**
  * End current write routine
@@ -152,6 +152,6 @@ size_t audiobuffer_write_fill(audiobuffer_t* b, const size_t frames);
  * @param b Pointer to initialized buffer
  * @return Number of frames was proccessed by current write routine
  */
-size_t audiobuffer_write_end(audiobuffer_t* b);
+uint32_t audiobuffer_write_end(audiobuffer_t* b);
 
 #endif
